refactor(client): Extract reconstructRelationships 🛠
